Fixed a variable definition bug in modelvalidation which lead to segmentation fault in arm architecture.

+++ src/modelvalidation.c
@@ -33,7 +33,12 @@ void random_kfold_group_generator(matrix **gid, size_t ngroups, size_t nobj, uns
 
 void kfold_group_train_test_split(matrix *x, matrix *y, matrix *gid, size_t group_id, matrix **x_train, matrix **y_train, matrix **x_test, matrix **y_test)
 {
-  /* Estimate how many objects are utilised to build the model and how many to predict */
+  /* Estimate how many objects are utilised to build the model and how many to predict
+   * gid is the group id matrix. Each row correspod to a group and each column to 
+   * an object id. This object id can be negative in the last line of the gid matrix
+   * which signify the end of the matrix.
+   *
+   */
   size_t i, j, k, l, n;
   size_t m_obj = 0;
   size_t p_obj = 0;
@@ -68,7 +73,7 @@ void kfold_group_train_test_split(matrix *x, matrix *y, matrix *gid, size_t grou
   for(i = 0, k = 0, l = 0; i < gid->row; i++){
     if(i != group_id){
       for(j = 0; j < gid->col; j++){
-        size_t a =  (size_t)gid->data[i][j]; /* get the row index */
+        int a =  (int)gid->data[i][j]; /* get the row index */
         if(a != -1){
           for(n = 0; n < x->col; n++){
             (*x_train)->data[k][n] = x->data[a][n];
@@ -85,7 +90,7 @@ void kfold_group_train_test_split(matrix *x, matrix *y, matrix *gid, size_t grou
     }
     else{
       for(j = 0; j < gid->col; j++){
-        size_t a = (size_t)gid->data[i][j];
+        int a = (int)gid->data[i][j];
         if(a != -1){
           for(n = 0; n < x->col; n++){
             (*x_test)->data[l][n] = x->data[a][n];
@@ -203,7 +208,7 @@ void *PLSRandomGroupCVModel(void *arg_)
     PLSYPredictorAllLV(x_test, subm, NULL, &y_test_predicted);
 
     for(j = 0, k = 0; j < gid->col; j++){
-      size_t a = (size_t)gid->data[g][j]; /*object id*/
+      int a = (int)gid->data[g][j]; /*object id*/
       if(a != -1){
         arg->predictioncounter->data[a] += 1; /* this object was visited */
         /* updating y */
@@ -274,7 +279,7 @@ void *MLRRandomGroupCVModel(void *arg_)
     MLRPredictY(x_test, NULL, subm, &y_test_predicted, NULL, NULL, NULL);
 
     for(j = 0, k = 0; j < gid->col; j++){
-      size_t a = (size_t)gid->data[g][j]; /*object id*/
+      int a = (int)gid->data[g][j]; /*object id*/
       if(a != -1){
         arg->predictioncounter->data[a] += 1; /* this object was visited */
         /* updating y */
@@ -343,7 +348,7 @@ void *EPLSRandomGroupCVModel(void *arg_)
     EPLSYPRedictorAllLV(x_test, subm, arg->crule, NULL, &y_test_predicted);
 
     for(j = 0, k = 0; j < gid->col; j++){
-      size_t a = (size_t)gid->data[g][j]; /*object id*/
+      int a = (int)gid->data[g][j]; /*object id*/
       if(a != -1){
         arg->predictioncounter->data[a] += 1; /* this object was visited */
         /* updating y */
